Workarounds for a crippled org.gnome.shell.Eval
Gnome Shell 41+ removes initial support for Eval. This means curtains
for the bootstrap code of gnome-shell-mode.

Eval can presently be enabled by an extension. One could make a
gnome-shell-mode-bootstrap-extension and enable that. The present
approach is to enable Eval via gnome-shell-extension and enable the
extension initially (assuming it is intalled in
$HOME/.local/share/gnome-shell-extensions.)

* extension.js: bring back org.gnome.Shell.Eval with
MetaContext.unsafe-mode=true.

* session.sh: start the gnome-shell session with gnome-shell-extension
as an installed and enabled extension. Also disable the version
verification check through dconf